add flags and fsanitize the code is workin perfect '

+++ Mandatory/lib/get_next_line.h
@@ -28,7 +28,7 @@ int		ft_strlen(char *s);
 
 void	*ft_calloc(size_t count, size_t size);
 
-char	*ft_substr(char *s, unsigned int start, size_t len);
+char	*ft_substr(char *s, int start, int len);
 
 char	*ft_strchr(char *s, int c);
